backend/requirements_old.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-JWT-Extended==4.5.3
yfinance==0.2.28
pandas==2.1.1
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
textblob==0.17.1
scikit-learn==1.3.0
tensorflow==2.13.0
websocket-client==1.6.3